Improve the robustness of the server side event processing
Now the events are being processed in the order of their kind:
1. Processing the joins, sending out all the necessary greeting and then
notifying players about the joiners.
2. We process all the players who have left.
3. Notify all the movements made by players.

This ensures that the events are processed in the correct order and we
only notify other players if the players stay for atleast a tick.

We also updated the "Hello" message so that players get all the
information about themselves as soon as they join instead of waiting for
aanother message to get information about their position and style. This
makes it so that we don't send another "PlayerJoined" event. Reducing
one event that needs to be sent.